+++ test/browser/karma-standalone.conf.js
@@ -11,6 +11,7 @@ module.exports = function (config) {
     basePath: '.',
     frameworks: ['mocha'],
     files: [
+      {pattern: 'vendor/polyfill.js', watch: false, included: true},
       {pattern: 'sway-standalone.js', watch: false, included: true},
       {pattern: 'test-browser.js', watch: false, included: true},
       {pattern: 'documents/**/*', watched: false, included: false, served: true}